chore: 收尾任务 T14/T15/T16/T17/T18
- T16: variant.price_delta Float->Numeric(12,2), 新增迁移0014
- T17: 改进 split_api.cjs 精确生成类型 import, 清理前端30处未用 import/变量, eslint 警告 1776->0, tsc 全绿
- T15: 新增并发下单压测脚本 loadtest_checkout.py (本地签发JWT绕过限流, Barrier同步检测超卖)
- T18: 修复56列仅依赖运行时ALTER兜底的真实缺陷, 新增迁移0015补齐列; 新增纯Alembic建库验收测试 test_migration_schema.py
- T14: 新增金额计算边界单测 test_order_amounts.py (compute_discount/get_tier/优惠聚合兜底/积分抵扣上限)
- 清理临时产物(_server.*、lint_stats.cjs、误操作文件)
- 后端 pytest 144 passed, 前端 tsc/eslint/build 全通过

+++ backend/app/models/variant.py
@@ -2,7 +2,7 @@ import json
 import uuid
 from datetime import datetime, timezone
 
-from sqlalchemy import Column, DateTime, Float, ForeignKey, Integer, String, Text
+from sqlalchemy import Column, DateTime, ForeignKey, Integer, Numeric, String, Text
 from sqlalchemy.orm import relationship
 
 from app.db.base import Base
@@ -15,7 +15,8 @@ class ProductVariant(Base):
     product_id = Column(String(36), ForeignKey("products.id"), nullable=False, index=True)
     sku_code = Column(String(60), nullable=True)
     specs = Column(Text, nullable=False, default="{}")  # JSON: {"颜色": "红", "尺寸": "L"}
-    price_delta = Column(Float, default=0.0)
+    # T16：参与订单金额计算，使用 Numeric(12,2) 精确到分
+    price_delta = Column(Numeric(12, 2), default=0)
     stock = Column(Integer, default=0)
     image_url = Column(String(512), nullable=True)
     created_at = Column(DateTime(timezone=True), default=lambda: datetime.now(timezone.utc))
